Add constrained exec_imports
I believe I have overloaded import key for constrained and unconstained working.

By this I mean "one ^ means constrained, two ^^ means unconstained".
The constrained will check "hey if we happen to not see a . or an alphabetic symbol after eating the import key, then that's an error".

We pass that error with a severity of zero.
Then the caller just ignores that error and continues.